Adds option to allow compasses to be offhanded. Refactors config file processes.
