+++ resources/modinfo.json
@@ -4,9 +4,10 @@
   "modid": "compass",
   "name": "Compass",
   "authors": [
-    "goxmeor"
+    "goxmeor",
+    "JapanHasRice"
   ],
   "description": "Adds craftable compasses",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "dependency": {}
 }
